updated weight entry thing

+++ model/home.go
@@ -1,6 +1,6 @@
 package model
 
 type HomePageData struct {
-	CurrentWeight float64
-	TargetWeight  float64
+	CurrentWeight int
+	TargetWeight  int
 }
