+++ model/weight.go
@@ -1,9 +1,9 @@
 package model
 
 type WeightEntry struct {
-	Id        int     `json:"id"`
-	UserId    int     `json:"userId"`
-	Weight    float64 `json:"weight"`
-	EntryDate string  `json:"entryDate"`
-	Category  string  `json:"category"`
+	Id        int    `json:"id"`
+	UserId    int    `json:"userId"`
+	Weight    int    `json:"weight"`
+	EntryDate string `json:"entryDate"`
+	Category  string `json:"category"`
 }
